--- frontend/src/scenes/actions/Actions.js
@@ -1,8 +1,0 @@
-import React from 'react'
-import { ActionsTable } from './ActionsTable'
-import { hot } from 'react-hot-loader/root'
-
-function _Actions(props) {
-    return <ActionsTable {...props} />
-}
-export const Actions = hot(_Actions)
