+++ frontend/src/scenes/actions/Actions.scss
@@ -2,7 +2,7 @@
 .tutorial-container {
     color: #c4c4c4;
     margin-top: 32px;
-    margin-bottom: 60px;
+    margin-bottom: 2rem;
     display: flex;
 }
 .t-element {
